Fixed a bad comparison error
Signed-off-by: Ritwik Gupta <ritwikg2004@live.com>

Initially I was starting with a `datetime.datetime` object, let's call it `a`. I was comparing it to a `datetime.date` object called `b` by doing `a.date() == b`. `a.date()` returns a datetime.date object.

However, later on I was assigning `a` to be a `datetime.date` object, which caused `a.date()` to error out.

I am now starting `a` as a `datetime.date` object, so this shouldn't be a problem anymore